chore: create a Progress component
